requirements.txt
fastapi
fizz
pydantic
pytest
PyPDF2
python-docx
pandas
openpyxl
langchain
langchain-openai
langgraph
openai
tiktoken
pytesseract
Pillow
pdf2image
python-dotenv
uvicorn
numpy
PyMuPDF
python-multipart
